// src/types/interfaces.d.ts
export declare enum SessionTypes {
    softphone = "softphone",
    collaborateVideo = "collaborateVideo",
    acdScreenShare = "screenShare",
    screenRecording = "screenRecording",
    liveScreenMonitoring = "liveScreenMonitoring",
    unknown = "unknown"
}
// ...
export type GenesysWebrtcBaseParams = {
    sessionId: string;
};
export type GenesysWebrtcSdpParams = GenesysWebrtcBaseParams & {
    sdp: string;
};
export type GenesysWebrtcOfferParams = GenesysWebrtcSdpParams & {
    conversationId: string;
    reinvite?: boolean;
};
export type GenesysInfoActiveParams = GenesysWebrtcBaseParams & {
    status: 'active';
};
export type GenesysSessionTerminateParams = GenesysWebrtcBaseParams & {
    reason?: JingleReasonCondition;
};
export type GenesysWebrtcMuteParams = GenesysWebrtcBaseParams & {
    type: 'audio' | 'video';
};
export type TypedJsonRpcMessage<Method extends string, Params> = {
    jsonrpc: string;
    method: Method;
    id?: string;
    params?: Params;
};
export type JsonRpcMessage = TypedJsonRpcMessage<string, any>;
export type GenesysWebrtcOffer = TypedJsonRpcMessage<'offer', GenesysWebrtcOfferParams>;
export type GenesysWebrtcAnswer = TypedJsonRpcMessage<'answer', GenesysWebrtcSdpParams>;
export type GenesysWebrtcInfo = TypedJsonRpcMessage<'info', GenesysInfoActiveParams>;
export type GenesysWebrtcIceCandidate = TypedJsonRpcMessage<'iceCandidate', GenesysWebrtcSdpParams>;
export type GenesysWebrtcTerminate = TypedJsonRpcMessage<'terminate', GenesysSessionTerminateParams>;
export type GenesysWebrtcMute = TypedJsonRpcMessage<'mute', GenesysWebrtcMuteParams>;
export type GenesysWebrtcUnmute = TypedJsonRpcMessage<'unmute', GenesysWebrtcMuteParams>;
export type GenesysWebrtcJsonRpcMessage = GenesysWebrtcOffer | GenesysWebrtcAnswer | GenesysWebrtcInfo | GenesysWebrtcIceCandidate | GenesysWebrtcTerminate | GenesysWebrtcMute | GenesysWebrtcUnmute;